a few more linux/gtk renames
